update unit test of map 24.11.22

+++ lib/algorithm.hpp
@@ -103,6 +103,7 @@ inline void print_space(int n) {
         std::cout << " "; 
     } 
 }
+
 template <typename Sequence>
 inline void print_heap_struct(const Sequence& seq) {
     int rounded_size = roundup22(seq.size());
@@ -139,9 +140,9 @@ inline void print_rbtree_heap_struct(const Sequence& seq) {
         if (node == nullptr) {
             std::cout << "\033[0;32m" << "." << "\033[0m";
         } else if (node->color) {
-            std::cout << "\033[0;32m" << seq[i]->val << "\033[0m";
+            std::cout << "\033[0;32m" << (int)seq[i]->val << "\033[0m";
         } else {
-            std::cout << "\033[0;31m" << seq[i]->val << "\033[0m";
+            std::cout << "\033[0;31m" << (int)seq[i]->val << "\033[0m";
         }
         
         print_space(max_width / col_num - 1);
@@ -155,7 +156,36 @@ inline void print_rbtree_heap_struct(const Sequence& seq) {
     }
     if (col_idx) { std::cout << "\n"; }
 }
+// template <typename Sequence>
+// inline void print_drbtree_heap_struct(const Sequence& seq) {
+//     int rounded_size = roundup22(seq.size());
+//     int max_width = rounded_size;
+//     int line_idx = 0;
+//     int col_num = 1;
+//     int col_idx = 0;
 
+//     for (int i = 0; i < seq.size(); i++) {
+//         print_space(max_width / col_num);
+//         auto node = seq[i];
+//         if (node == nullptr) {
+//             std::cout << "\033[0;32m" << "." << "\033[0m";
+//         } else if (node->color) {
+//             std::cout << "\033[0;32m" << seq[i]->val.second << "\033[0m";
+//         } else {
+//             std::cout << "\033[0;31m" << seq[i]->val.second << "\033[0m";
+//         }
+        
+//         print_space(max_width / col_num - 1);
+//         col_idx++;
+//         if (col_idx >= col_num) {
+//             line_idx++;
+//             col_num *= 2;
+//             col_idx = 0;
+//             std::cout << "\n";
+//         }
+//     }
+//     if (col_idx) { std::cout << "\n"; }
+// }
 
 template <typename ForwardIt, 
           typename T = typename mfwu::iterator_traits<ForwardIt>::value_type,
